DIG-30 * make marker set location upon drag
