+++ modules/screens/test/all-tests/LocationStep.js
@@ -2,23 +2,24 @@ import StyledText from '../../../components/StyledText/StyledText';
 import { VStack } from 'native-base';
 import MapView, { Marker } from 'react-native-maps';
 import { StyleSheet } from 'react-native';
+import PropTypes from 'prop-types';
 
-const initialRegion = {
-    latitude: 49.2827,
-    longitude: -123.116226,
-    latitudeDelta: 0.0922,
-    longitudeDelta: 0.0421,
-};
+const latitudeDelta = 0.0922;
+const longitudeDelta = 0.0421;
 
-export default function LocationStep() {
+export default function LocationStep({ setLocation, location }) {
     return (
         <VStack alignItems="center" space={20} marginTop="30px">
             <StyledText>Location Step</StyledText>
-            <MapView style={styles.map} initialRegion={initialRegion}>
+            <MapView
+                style={styles.map}
+                initialRegion={{ ...location, longitudeDelta, latitudeDelta }}
+            >
                 <Marker
-                    coordinate={{
-                        latitude: initialRegion.latitude,
-                        longitude: initialRegion.longitude,
+                    coordinate={location}
+                    draggable={true}
+                    onDragEnd={(e) => {
+                        setLocation(e.nativeEvent.coordinate);
                     }}
                 ></Marker>
             </MapView>
@@ -32,3 +33,8 @@ const styles = StyleSheet.create({
         height: '100%',
     },
 });
+
+LocationStep.propTypes = {
+    setLocation: PropTypes.func.isRequired,
+    location: PropTypes.shape({}).isRequired,
+};
