+++ modules/screens/test/compression-test/CompressionTestWizard.js
@@ -8,6 +8,11 @@ import SnowConditionTestStep from '../all-tests/SnowConditionTestStep';
 import TitleTestStep from '../all-tests/TitleTestStep';
 import LocationStep from '../all-tests/LocationStep';
 
+const initialRegion = {
+    latitude: 49.2827,
+    longitude: -123.116226,
+};
+
 export default function CompressionTestWizard(props) {
     const [testData, setTestData] = useState({
         tapResult: '',
@@ -16,7 +21,10 @@ export default function CompressionTestWizard(props) {
         weather: '',
         snowCondition: '',
         title: '',
-        location: '',
+        location: {
+            latitude: initialRegion.latitude,
+            longitude: initialRegion.longitude,
+        },
     });
 
     const updateByKey = (key) => (value) => {
